arch/x86_64: Fix return type of ffsl and flsl
The commit fixes the return type for ffsl and flsl which should return
an unsigned integer like ffs and fls. The promotion to unsigned long is
unnecessary.

+++ arch/x86/x86_64/include/uk/asm/atomic.h
@@ -67,7 +67,7 @@ static inline unsigned int ukarch_fls(unsigned int x)
  * @return The index of the least significant 1-bit of x, or if x is zero,
  *    the result is undefined. ffs(1)=0, ffs(3)=0, ffs(0x80000000)=31
  */
-static inline unsigned long ukarch_ffsl(unsigned long x)
+static inline unsigned int ukarch_ffsl(unsigned long x)
 {
 	__asm__("bsfq %1,%0"
 		: "=r" (x)
@@ -82,7 +82,7 @@ static inline unsigned long ukarch_ffsl(unsigned long x)
  * @return The index of the most significant 1-bit of x, or if x is zero,
  *   the result is undefined. fls(1)=0, fls(3)=1, fls(0x80000001)=31
  */
-static inline unsigned long ukarch_flsl(unsigned long x)
+static inline unsigned int ukarch_flsl(unsigned long x)
 {
 	__asm__("bsrq %1,%0"
 		: "=r" (x)
